fix: avoid tautological comparison warning when checking ncpu range
Remove unnecessary cast to (long)INT_MAX in the range check for ncpu.
The cast could trigger -Wtautological-type-limit-compare on platforms
where long is wider than int, making the comparison appear always true.

Use direct comparison against INT_MAX instead, relying on implicit
type promotion. This eliminates the warning while preserving the
intended bounds check.

+++ src/util.c
@@ -268,7 +268,7 @@ int get_ncpu(void) {
             ncpu = get_online_cpu_count();
         }
 #endif
-        cached_ncpu = (ncpu > 0 && ncpu <= (long)INT_MAX) ? (int)ncpu : 1;
+        cached_ncpu = (ncpu > 0 && ncpu <= INT_MAX) ? (int)ncpu : 1;
 
 #elif defined(__APPLE__) || defined(__FreeBSD__)
         /* macOS and FreeBSD: use sysctl interface */
